+++ src/Smartphone.java
@@ -5,13 +5,13 @@ public class Smartphone extends Device {
     private String model;
     private String description;
     private double displaySize;
-    private double storageSize;
+    private int storageSize;
     private double purchasePrice;
     private double sellingPrice;
     private int deviceId;
 
     public Smartphone (String deviceType, String manufacturer, String model, String description,
-                       double displaySize, double storageSize, double purchasePrice, double sellingPrice,
+                       double displaySize, int storageSize, double purchasePrice, double sellingPrice,
                        int deviceId){
         this.deviceType= deviceType;
         this.manufacturer= manufacturer;
@@ -44,8 +44,8 @@ public class Smartphone extends Device {
         return displaySize;
     }
 
-    public double getStorageSize() {
-        return (int) storageSize;
+    public int getStorageSize() {
+        return storageSize;
     }
 
     public double getPurchasePrice() {
@@ -80,7 +80,7 @@ public class Smartphone extends Device {
         this.displaySize = displaySize;
     }
 
-    public void setStorageSize(double storageSize) {
+    public void setStorageSize(int storageSize) {
         this.storageSize = storageSize;
     }
